refactor(db): extract covers I/O to db/src/covers.rs
